requirements.txt
click==8.1.7
cmake==3.26.4
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.4
opencv-python==4.8.0.74
pillow==10.3.0
setuptools==70.0.0
mysql-connector-python==8.3.0click==8.1.7
cmake==3.26.4
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.8.0.74
pillow==10.3.0
python-dotenv==1.0.1
setuptools==70.0.0